+++ test/smoke/src/test-helpers/index.ts
@@ -1,0 +1,8 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (C) 2024 Posit Software, PBC. All rights reserved.
+ *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+export { prepareTestEnv } from './testSetup';
+export { cloneTestRepo } from './utils';
+export { runMochaTests } from './mochaRunner';
